SCM API routes
